+++ contracts/payment/escrow/Escrow.sol
@@ -1,14 +1,19 @@
 pragma solidity ^0.4.24;
 
-import "../math/SafeMath.sol";
-import "../ownership/Secondary.sol";
+import "../../math/SafeMath.sol";
+import "../../ownership/Secondary.sol";
 
-/**
+ /**
  * @title Escrow
- * @dev Base escrow contract, holds funds destinated to a payee until they
- * withdraw them. The contract that uses the escrow as its payment method
- * should be its primary, and provide public methods redirecting to the escrow's
- * deposit and withdraw.
+ * @dev Base escrow contract, holds funds designated for a payee until they
+ * withdraw them.
+ * @dev Intended usage: This contract (and derived escrow contracts) should be a
+ * standalone contract, that only interacts with the contract that instantiated
+ * it. That way, it is guaranteed that all Ether will be handled according to
+ * the Escrow rules, and there is no need to check for payable functions or
+ * transfers in the inheritance tree. The contract that uses the escrow as its
+ * payment method should be its primary, and provide public methods redirecting
+ * to the escrow's deposit and withdraw.
  */
 contract Escrow is Secondary {
   using SafeMath for uint256;
