+++ contracts/payment/escrow/RefundEscrow.sol
@@ -4,9 +4,14 @@ import "./ConditionalEscrow.sol";
 
 /**
  * @title RefundEscrow
- * @dev Escrow that holds funds for a beneficiary, deposited from multiple parties.
- * The primary account may close the deposit period, and allow for either withdrawal
- * by the beneficiary, or refunds to the depositors.
+ * @dev Escrow that holds funds for a beneficiary, deposited from multiple
+ * parties.
+ * @dev Intended usage: See Escrow.sol. Same usage guidelines apply here.
+ * @dev The primary account (that is, the contract that instantiates this
+ * contract) may deposit, close the deposit period, and allow for either
+ * withdrawal by the beneficiary, or refunds to the depositors. All interactions
+ * with RefundEscrow will be made through the primary contract. See the
+ * RefundableCrowdsale contract for an example of RefundEscrow’s use.
  */
 contract RefundEscrow is ConditionalEscrow {
   enum State { Active, Refunding, Closed }
