+++ test/payment/escrow/RefundEscrow.test.js
@@ -1,8 +1,8 @@
-const shouldFail = require('../helpers/shouldFail');
-const expectEvent = require('../helpers/expectEvent');
-const { ethGetBalance } = require('../helpers/web3');
-const { ether } = require('../helpers/ether');
-const { ZERO_ADDRESS } = require('../helpers/constants');
+const shouldFail = require('../../helpers/shouldFail');
+const expectEvent = require('../../helpers/expectEvent');
+const { ethGetBalance } = require('../../helpers/web3');
+const { ether } = require('../../helpers/ether');
+const { ZERO_ADDRESS } = require('../../helpers/constants');
 
 const BigNumber = web3.BigNumber;
